OLMIS-4105 Fix UI getting status message
The UI needs to look for the locale key that is a correct locale string. The approach here is done by replacing hyphens that come from the APIs, with underscores that match what the HTML is expecting. This is done because it is unclear how to use Angular interpolation with JavaScript object bracket notation ({{ obj['key'] }})?